+++ Reach/src/utopia/reach/component/interactive/input/selection/SelectableStack.scala
@@ -1,21 +1,27 @@
 package utopia.reach.component.interactive.input.selection
 
 import utopia.firmament.component.input.{InputWithPointer, SelectionWithPointers}
+import utopia.firmament.context.ComponentCreationDefaults
 import utopia.firmament.context.color.VariableColorContextLike
+import utopia.firmament.context.text.VariableTextContext
 import utopia.firmament.controller.StackItemAreas
 import utopia.firmament.controller.data.SelectionKeyListener
 import utopia.firmament.drawing.template.CustomDrawer
 import utopia.firmament.model.GuiElementStatus
 import utopia.firmament.model.enumeration.GuiElementState.Focused
 import utopia.firmament.model.enumeration.MouseInteractionState.{Hover, Pressed}
+import utopia.firmament.model.enumeration.StackLayout
+import utopia.firmament.model.stack.StackLength
 import utopia.flow.collection.CollectionExtensions._
-import utopia.flow.collection.immutable.Pair
+import utopia.flow.collection.immutable.{Empty, Pair}
 import utopia.flow.event.listener.ChangeListener
 import utopia.flow.operator.equality.EqualsFunction
 import utopia.flow.operator.filter.{AcceptAll, Filter}
 import utopia.flow.operator.sign.Sign
 import utopia.flow.operator.sign.Sign.{Negative, Positive}
-import utopia.flow.view.immutable.eventful.{AlwaysFalse, LazilyInitializedChanging}
+import utopia.flow.util.Mutate
+import utopia.flow.util.logging.Logger
+import utopia.flow.view.immutable.eventful.{AlwaysFalse, AlwaysTrue, LazilyInitializedChanging}
 import utopia.flow.view.mutable.Pointer
 import utopia.flow.view.mutable.eventful.EventfulPointer
 import utopia.flow.view.template.eventful.{Changing, Flag}
@@ -25,16 +31,385 @@ import utopia.genesis.handling.event.consume.ConsumeChoice
 import utopia.genesis.handling.event.keyboard.Key.{LeftArrow, RightArrow, UpArrow}
 import utopia.genesis.handling.event.keyboard.{Key, KeyboardEvents}
 import utopia.genesis.handling.event.mouse._
+import utopia.paradigm.color.ColorLevel.Standard
+import utopia.paradigm.color.{ColorLevel, ColorRole}
 import utopia.paradigm.enumeration.Axis.{X, Y}
+import utopia.paradigm.enumeration.Axis2D
 import utopia.paradigm.shape.shape2d.area.polygon.c4.bounds.Bounds
 import utopia.paradigm.shape.shape2d.vector.point.Point
+import utopia.reach.component.factory.FocusListenableFactory
 import utopia.reach.component.factory.FromContextComponentFactoryFactory.Ccff
 import utopia.reach.component.factory.contextual.ContextualFactory
 import utopia.reach.component.hierarchy.ComponentHierarchy
 import utopia.reach.component.template.focus.FocusableWithState
-import utopia.reach.component.template.{HasGuiState, ReachComponent, ReachComponentWrapper}
-import utopia.reach.container.multi.{ViewStack, ViewStackSettings}
+import utopia.reach.component.template.{HasGuiState, PartOfComponentHierarchy, ReachComponent, ReachComponentWrapper}
+import utopia.reach.container.multi.{SegmentGroup, ViewStack, ViewStackSettings, ViewStackSettingsLike}
 import utopia.reach.focus.{FocusListener, FocusStateTracker}
+
+/**
+ * Common trait for selectable stack factories and settings
+ * @tparam Repr Implementing factory/settings type
+ * @author Mikko Hilpinen
+ * @since 12.09.2025, v1.6
+ */
+trait SelectableStackSettingsLike[+Repr] extends FocusListenableFactory[Repr] with ViewStackSettingsLike[Repr]
+{
+	// ABSTRACT	--------------------
+	
+	/**
+	 * Settings that affect this stack's layout
+	 */
+	def stackSettings: ViewStackSettings
+	
+	/**
+	 * A drawer used for visualizing selection and mouse interaction
+	 */
+	def selectionDrawer: Option[SelectionDrawer]
+	/**
+	 * Keys (other than the arrow keys), which are used for moving the selection around.
+	 * Each key is mapped to the direction to which it moves the selection.
+	 */
+	def extraSelectionKeys: Map[Key, Sign]
+	/**
+	 * A flag that, when set, makes the keyboard-based selection function even when the component
+	 * doesn't have focus
+	 */
+	def alternativeKeySelectionEnabledFlag: Flag
+	/**
+	 * Whether arrow key -based selection should be enabled
+	 */
+	def arrowKeySelectionEnabled: Boolean
+	
+	/**
+	 * Changes the keyboard keys used for moving the selection around.
+	 * Note: Overrides previously used keys, including the default arrow keys.
+	 * @param keys Keys which are used for moving the selection around.
+	 *             Each key is mapped to the direction to which it moves the selection.
+	 * @return Copy of this factory with the specified keys used for selection (exclusively)
+	 */
+	def withSelectionKeys(keys: Map[Key, Sign]): Repr
+	/**
+	 * Keys (other than the arrow keys), which are used for moving the selection around.
+	 * Each key is mapped to the direction to which it moves the selection.
+	 * @param keys New additional selection keys to use.
+	 *             Keys (other than the arrow keys), which are used for moving the selection around.
+	 *             Each key is mapped to the direction to which it moves the selection.
+	 * @return Copy of this factory with the specified additional selection keys
+	 */
+	def withExtraSelectionKeys(keys: Map[Key, Sign]): Repr
+	/**
+	 * A flag that, when set, makes the keyboard-based selection function even when the component
+	 * doesn't have focus
+	 * @param enabledFlag New alternative key selection enabled flag to use.
+	 *                    A flag that, when set, makes the keyboard-based selection function even
+	 *                    when the component doesn't have focus
+	 * @return Copy of this factory with the specified alternative key selection enabled flag
+	 */
+	def withAlternativeKeySelectionEnabledFlag(enabledFlag: Flag): Repr
+	/**
+	 * Whether arrow key -based selection should be enabled
+	 * @param enabled New arrow key selection enabled to use.
+	 *                Whether arrow key -based selection should be enabled
+	 * @return Copy of this factory with the specified arrow key selection enabled
+	 */
+	def withArrowKeySelectionEnabled(enabled: Boolean): Repr
+	/**
+	 * A drawer used for visualizing selection and mouse interaction
+	 * @param drawer New selection drawer to use.
+	 *               A drawer used for visualizing selection and mouse interaction
+	 * @return Copy of this factory with the specified selection drawer
+	 */
+	def withSelectionDrawer(drawer: Option[SelectionDrawer]): Repr
+	
+	/**
+	 * Settings that affect this stack's layout
+	 * @param settings New stack settings to use.
+	 *                 Settings that affect this stack's layout
+	 * @return Copy of this factory with the specified stack settings
+	 */
+	def withStackSettings(settings: ViewStackSettings): Repr
+	
+	
+	// COMPUTED ------------------------
+	
+	/**
+	 * @return Copy of this factory without a selection drawer.
+	 *         Useful when the generated components visualize their selection state independently.
+	 */
+	def withoutSelectionDrawer = withSelectionDrawer(None)
+	
+	/**
+	 * @return Copy of this factory where the created stacks react to keyboard events even while they're not in focus.
+	 */
+	def withoutFocusKeyRequirement = withAlternativeKeySelectionEnabledFlag(AlwaysTrue)
+	
+	
+	// IMPLEMENTED	--------------------
+	
+	override def axisPointer: Changing[Axis2D] = stackSettings.axisPointer
+	override def layoutPointer: Changing[StackLayout] = stackSettings.layoutPointer
+	override def segmentGroup: Option[SegmentGroup] = stackSettings.segmentGroup
+	override def capPointer: Changing[StackLength] = stackSettings.capPointer
+	override def customDrawers = stackSettings.customDrawers
+	
+	override def withAxisPointer(p: Changing[Axis2D]): Repr = mapStackSettings { _.withAxisPointer(p) }
+	override def withLayoutPointer(p: Changing[StackLayout]): Repr = mapStackSettings { _.withLayoutPointer(p) }
+	override def withSegmentGroup(group: Option[SegmentGroup]): Repr = mapStackSettings { _.withSegmentGroup(group) }
+	override def withCapPointer(p: Changing[StackLength]): Repr = mapStackSettings { _.withCapPointer(p) }
+	override def withCustomDrawers(drawers: Seq[CustomDrawer]) = mapStackSettings { _.withCustomDrawers(drawers) }
+	
+	
+	// OTHER	--------------------
+	
+	/**
+	 * A drawer used for visualizing selection and mouse interaction
+	 * @param drawer New selection drawer to use.
+	 *               A drawer used for visualizing selection and mouse interaction
+	 * @return Copy of this factory with the specified selection drawer
+	 */
+	def withSelectionDrawer(drawer: SelectionDrawer): Repr = withSelectionDrawer(Some(drawer))
+	
+	/**
+	 * Changes the keyboard keys used for moving the selection around.
+	 * Note: Overrides previously used keys, including the default arrow keys.
+	 * @param backKey Key used for moving the selection backward
+	 * @param forwardKey Key used for moving the selection forward
+	 * @return Copy of this factory with the specified keys used for selection (exclusively)
+	 */
+	def withSelectionKeys(backKey: Key, forwardKey: Key): Repr =
+		withSelectionKeys(Map(backKey -> Negative, forwardKey -> Positive))
+	
+	def withAdditionalSelectionKeys(backKey: Key, forwardKey: Key): Repr =
+		withAdditionalSelectionKeys(Map(backKey -> Negative, forwardKey -> Positive))
+	def withAdditionalSelectionKeys(keys: Map[Key, Sign]): Repr = mapExtraSelectionKeys { _ ++ keys }
+	
+	/**
+	 * @param backKey Key used for moving the selection backward
+	 * @param forwardKey Key used for moving the selection forward
+	 * @param exclusive Whether these should be the only keys used for moving the selection (default = false)
+	 * @return Copy of this factory with the specified selection keys
+	 */
+	def movingSelectionUsing(backKey: Key, forwardKey: Key, exclusive: Boolean = false) = {
+		if (exclusive)
+			withSelectionKeys(backKey, forwardKey)
+		else
+			withAdditionalSelectionKeys(backKey, forwardKey)
+	}
+	
+	def mapExtraSelectionKeys(f: Mutate[Map[Key, Sign]]) = withExtraSelectionKeys(f(extraSelectionKeys))
+	def mapAlternativeKeySelectionEnabledFlag(f: Mutate[Flag]) =
+		withAlternativeKeySelectionEnabledFlag(f(alternativeKeySelectionEnabledFlag))
+	
+	def mapStackSettings(f: Mutate[ViewStackSettings]) = withStackSettings(f(stackSettings))
+}
+
+object SelectableStackSettings
+{
+	// ATTRIBUTES	--------------------
+	
+	val default = apply()
+}
+/**
+ * Combined settings used when constructing selectable stacks
+ * @param focusListeners                     Focus listeners to assign to created components
+ * @param stackSettings                      Settings that affect this stack's layout
+ * @param selectionDrawer                    A drawer used for visualizing selection and mouse
+ *                                           interaction
+ * @param extraSelectionKeys            Keys (other than the arrow keys), which are used
+ *                                           for moving the selection around.
+ *                                           Each key is mapped to the direction to which it
+ *                                           moves the selection.
+ * @param alternativeKeySelectionEnabledFlag A flag that, when set, makes the keyboard-based
+ *                                           selection function even when the component doesn't
+ *                                           have focus
+ * @param arrowKeySelectionEnabled           Whether arrow key -based selection should be enabled
+ * @author Mikko Hilpinen
+ * @since 12.09.2025, v1.6
+ */
+case class SelectableStackSettings(focusListeners: Seq[FocusListener] = Empty,
+                                   stackSettings: ViewStackSettings = ViewStackSettings.default,
+                                   selectionDrawer: Option[SelectionDrawer] = None,
+                                   extraSelectionKeys: Map[Key, Sign] = Map[Key, Sign](),
+                                   alternativeKeySelectionEnabledFlag: Flag = AlwaysFalse,
+                                   arrowKeySelectionEnabled: Boolean = true)
+	extends SelectableStackSettingsLike[SelectableStackSettings]
+{
+	// IMPLEMENTED	--------------------
+	
+	override def withExtraSelectionKeys(keys: Map[Key, Sign]) =
+		copy(extraSelectionKeys = keys)
+	override def withAlternativeKeySelectionEnabledFlag(enabledFlag: Flag) =
+		copy(alternativeKeySelectionEnabledFlag = enabledFlag)
+	override def withArrowKeySelectionEnabled(enabled: Boolean) =
+		copy(arrowKeySelectionEnabled = enabled)
+	override def withFocusListeners(listeners: Seq[FocusListener]) =
+		copy(focusListeners = listeners)
+	override def withSelectionDrawer(drawer: Option[SelectionDrawer]) =
+		copy(selectionDrawer = drawer)
+	override def withStackSettings(settings: ViewStackSettings) = copy(stackSettings = settings)
+	
+	override def withSelectionKeys(keys: Map[Key, Sign]): SelectableStackSettings =
+		copy(extraSelectionKeys = keys, arrowKeySelectionEnabled = false)
+}
+
+/**
+ * Common trait for factories that wrap a selectable stack settings instance
+ * @tparam Repr Implementing factory/settings type
+ * @author Mikko Hilpinen
+ * @since 12.09.2025, v1.6
+ */
+trait SelectableStackSettingsWrapper[+Repr] extends SelectableStackSettingsLike[Repr]
+{
+	// ABSTRACT	--------------------
+	
+	/**
+	 * Settings wrapped by this instance
+	 */
+	protected def settings: SelectableStackSettings
+	/**
+	 * @return Copy of this factory with the specified settings
+	 */
+	def withSettings(settings: SelectableStackSettings): Repr
+	
+	
+	// IMPLEMENTED	--------------------
+	
+	override def extraSelectionKeys = settings.extraSelectionKeys
+	override def alternativeKeySelectionEnabledFlag = settings.alternativeKeySelectionEnabledFlag
+	override def arrowKeySelectionEnabled = settings.arrowKeySelectionEnabled
+	override def focusListeners = settings.focusListeners
+	override def selectionDrawer = settings.selectionDrawer
+	override def stackSettings = settings.stackSettings
+	
+	override def withExtraSelectionKeys(keys: Map[Key, Sign]) =
+		mapSettings { _.withExtraSelectionKeys(keys) }
+	override def withAlternativeKeySelectionEnabledFlag(enabledFlag: Flag) =
+		mapSettings { _.withAlternativeKeySelectionEnabledFlag(enabledFlag) }
+	override def withArrowKeySelectionEnabled(enabled: Boolean) =
+		mapSettings { _.withArrowKeySelectionEnabled(enabled) }
+	override def withFocusListeners(listeners: Seq[FocusListener]) =
+		mapSettings { _.withFocusListeners(listeners) }
+	override def withSelectionDrawer(drawer: Option[SelectionDrawer]) =
+		mapSettings { _.withSelectionDrawer(drawer) }
+	override def withStackSettings(settings: ViewStackSettings) = mapSettings { _.withStackSettings(settings) }
+	
+	override def withSelectionKeys(keys: Map[Key, Sign]): Repr = mapSettings { _.withSelectionKeys(keys) }
+	
+	
+	// OTHER	--------------------
+	
+	def mapSettings(f: SelectableStackSettings => SelectableStackSettings) = withSettings(f(settings))
+}
+
+/**
+ * Factory class used for constructing selectable stacks using contextual component creation
+ * information
+ * @author Mikko Hilpinen
+ * @since 12.09.2025, v1.6
+ */
+case class SelectableStackFactory[N <: VariableColorContextLike[N, _]](hierarchy: ComponentHierarchy, context: N,
+                                                                       settings: SelectableStackSettings = SelectableStackSettings.default)
+	extends SelectableStackSettingsWrapper[SelectableStackFactory[N]] with PartOfComponentHierarchy
+{
+	// COMPUTED ------------------------
+	
+	/**
+	 * @return Copy of this factory using a selection drawer that highlights the selected area's background
+	 */
+	def highlightingSelectedArea =
+		withContextualSelectionDrawer { SelectionDrawer.highlight(_) }
+	
+	
+	// IMPLEMENTED	--------------------
+	
+	override def withSettings(settings: SelectableStackSettings) = copy(settings = settings)
+	
+	
+	// IMPLICIT ------------------------
+	
+	private implicit def log: Logger = ComponentCreationDefaults.componentLogger
+	
+	
+	// OTHER    ------------------------
+	
+	def withContext[N2 <: VariableColorContextLike[N2, _]](newContext: N2): SelectableStackFactory[N2] =
+		copy(context = newContext)
+	def mapContext[N2 <: VariableColorContextLike[N2, _]](f: N => N2) =
+		withContext(f(context))
+	
+	/**
+	 * @param color Selection background color
+	 * @param preferredShade Preferred background shade (default = standard)
+	 * @param highlightModifier A modifier applied to highlighting intensity.
+	 *                          Highlighting is used for visualizing the mouse hover & focus states.
+	 *                          Default = 1.0 = Original.
+	 * @return Copy of this factory highlighting the selected area using the specified background color
+	 */
+	def withSelectionBackground(color: ColorRole, preferredShade: ColorLevel = Standard,
+	                            highlightModifier: Double = 1.0) =
+		withContextualSelectionDrawer { SelectionDrawer.fill(_, color, preferredShade, highlightModifier) }
+	/**
+	 * @param createDrawer A function that receives this factory's context, and uses it to create a selection drawer
+	 * @return Copy of this factory using the constructed selection drawer
+	 */
+	def withContextualSelectionDrawer(createDrawer: N => SelectionDrawer) =
+		withSelectionDrawer(createDrawer(context))
+	
+	/**
+	 * Creates a new selectable stack
+	 * @param viewFactory A factory used for constructing the stacked components
+	 * @param contentPointer A pointer that contains the displayed content
+	 * @param valuePointer A mutable pointer that contains the selected value (default = new pointer)
+	 * @param makeView A function that receives:
+	 *                      1. Component factory
+	 *                      1. Content pointer
+	 *                      1. Selection flag
+	 *                      1. Component index (0-based)
+	 *
+	 *                 And yields a component to display at that position
+	 * @param eq Implicit equals-function used for comparing selection. Default = Use ==
+	 * @tparam A Type of displayed & selected values
+	 * @tparam F Type of the component factories used
+	 * @return A new selectable stack component
+	 */
+	def apply[A, F <: ContextualFactory[N, F]](viewFactory: Ccff[N, F], contentPointer: Changing[Seq[A]],
+	                                           valuePointer: EventfulPointer[Option[A]] = Pointer.eventful.empty[A])
+	                                          (makeView: (F, Changing[A], Flag, Int) => ReachComponent)
+	                                          (implicit eq: EqualsFunction[A] = EqualsFunction.default) =
+		new SelectableStack[A, N, F](hierarchy, context, contentPointer, valuePointer, viewFactory, makeView, settings)
+}
+
+/**
+ * Used for defining selectable stack creation settings outside the component building process
+ * @author Mikko Hilpinen
+ * @since 12.09.2025, v1.6
+ */
+case class SelectableStackSetup(settings: SelectableStackSettings = SelectableStackSettings.default)
+	extends SelectableStackSettingsWrapper[SelectableStackSetup]
+		// Only extends Ccff because the generic version requires a self-referencing type
+		with Ccff[VariableTextContext, SelectableStackFactory[VariableTextContext]]
+{
+	// IMPLEMENTED	--------------------
+	
+	override def withContext(hierarchy: ComponentHierarchy,
+	                         context: VariableTextContext): SelectableStackFactory[VariableTextContext] =
+		withContext[VariableTextContext](hierarchy, context)
+	
+	override def withSettings(settings: SelectableStackSettings) = copy(settings = settings)
+	
+	
+	// OTHER    -------------------------
+	
+	def withContext[N <: VariableColorContextLike[N, _]](hierarchy: ComponentHierarchy, context: N) =
+		SelectableStackFactory(hierarchy, context, settings)
+}
+
+object SelectableStack extends SelectableStackSetup()
+{
+	// OTHER	--------------------
+	
+	def apply(settings: SelectableStackSettings) = withSettings(settings)
+}
 
 /**
  * Adds selection features to a view stack
@@ -45,11 +420,9 @@ import utopia.reach.focus.{FocusListener, FocusStateTracker}
  * @since 09.09.2025, v1.7
  */
 class SelectableStack[A, N <: VariableColorContextLike[N, _], F <: ContextualFactory[N, F]]
-(override val hierarchy: ComponentHierarchy, context: N, stackSettings: ViewStackSettings,
- override val contentPointer: Changing[Seq[A]], override val valuePointer: EventfulPointer[Option[A]],
- viewFactory: Ccff[N, F], makeView: (F, Changing[A], Flag, Int) => ReachComponent,
- selectionDrawer: Option[SelectionDrawer], arrowKeySelectionEnabled: Boolean, otherSelectionKeys: Map[Key, Sign],
- alternativeKeySelectionCondition: Flag = AlwaysFalse, otherFocusListeners: Seq[FocusListener])
+(override val hierarchy: ComponentHierarchy, context: N, override val contentPointer: Changing[Seq[A]],
+ override val valuePointer: EventfulPointer[Option[A]], viewFactory: Ccff[N, F],
+ makeView: (F, Changing[A], Flag, Int) => ReachComponent, settings: SelectableStackSettings)
 (implicit eq: EqualsFunction[A])
 	extends ReachComponentWrapper with FocusableWithState with HasGuiState
 		with SelectionWithPointers[Option[A], EventfulPointer[Option[A]], Seq[A], Changing[Seq[A]]]
@@ -59,7 +432,7 @@ class SelectableStack[A, N <: VariableColorContextLike[N, _], F <: ContextualFac
 	
 	override lazy val focusId: Int = hashCode()
 	private val focusTracker = new FocusStateTracker()
-	override lazy val focusListeners: Seq[FocusListener] = focusTracker +: otherFocusListeners
+	override lazy val focusListeners: Seq[FocusListener] = focusTracker +: settings.focusListeners
 	override val allowsFocusLeave: Boolean = true
 	
 	private val stateP = Pointer.eventful(GuiElementStatus.identity)
@@ -68,14 +441,14 @@ class SelectableStack[A, N <: VariableColorContextLike[N, _], F <: ContextualFac
 		selected.flatMap { selected => content.findIndexWhere { eq(selected, _) } }
 	}
 	
-	override protected lazy val wrapped = ViewStack.withContext(hierarchy, context).withSettings(stackSettings)
-		.withCustomDrawer(selectionDrawer match {
+	override protected lazy val wrapped = ViewStack.withContext(hierarchy, context).withSettings(settings.stackSettings)
+		.withCustomDrawer(settings.selectionDrawer match {
 			case Some(drawer) => new LocalSelectionDrawer(drawer)
 			case None => new HoverDrawer
 		})
 		.mapPointer(contentPointer, viewFactory) { (factory, pointer, index) =>
 			// Tracks selection status and modifies the background pointer accordingly
-			val (selectedFlag, correctBgFactory) = selectionDrawer.flatMap { _.selectionBackgroundPointer } match {
+			val (selectedFlag, correctBgFactory) = settings.selectionDrawer.flatMap { _.selectionBackgroundPointer } match {
 				// Case: Selected item background is different from normal => Creates a custom background color -pointer
 				case Some(bgP) =>
 					val selectedFlag: Flag = selectedIndexP.lightMap { _.contains(index) }
@@ -94,7 +467,7 @@ class SelectableStack[A, N <: VariableColorContextLike[N, _], F <: ContextualFac
 	
 	private lazy val locationTracker = new StackItemAreas[ReachComponent](wrapped, componentsP)
 	
-	private val listensToKeyboard = arrowKeySelectionEnabled || otherSelectionKeys.nonEmpty
+	private val listensToKeyboard = settings.arrowKeySelectionEnabled || settings.extraSelectionKeys.nonEmpty
 	
 	/**
 	 * A change listener for repainting changed selection and/or hover areas
@@ -123,15 +496,16 @@ class SelectableStack[A, N <: VariableColorContextLike[N, _], F <: ContextualFac
 		
 		// Sets up key listening
 		if (listensToKeyboard) {
-			val keysP = stackSettings.axisPointer.map { axis =>
+			val keysP = settings.axisPointer.map { axis =>
 				val axisKeys: Map[Key, Sign] = axis match {
 					case X => Map(LeftArrow -> Negative, RightArrow -> Positive)
 					case Y => Map(UpArrow -> Negative, RightArrow -> Positive)
 				}
-				axisKeys ++ otherSelectionKeys
+				axisKeys ++ settings.extraSelectionKeys
 			}
 			val listener = SelectionKeyListener
-				.withEnabledFlag((linkedFlag && wrapped.visibleFlag) && (focusFlag || alternativeKeySelectionCondition))
+				.withEnabledFlag((linkedFlag && wrapped.visibleFlag) &&
+					(focusFlag || settings.alternativeKeySelectionEnabledFlag))
 				.listeningTo(keysP)
 				.apply(moveSelectionBy)
 			
@@ -189,6 +563,14 @@ class SelectableStack[A, N <: VariableColorContextLike[N, _], F <: ContextualFac
 				}
 		}
 	}
+	/**
+	 * Moves the selection to the next available item
+	 */
+	def selectNext() = moveSelectionBy(1)
+	/**
+	 * Moves the selection to the previous available item
+	 */
+	def selectPrevious() = moveSelectionBy(-1)
 	
 	
 	// NESTED   ---------------------------